+++ Rubberduck.Core/UI/Command/AboutCommand.cs
@@ -1,6 +1,7 @@
 ﻿using System.Runtime.InteropServices;
 using NLog;
 using Rubberduck.UI.About;
+using Rubberduck.VersionCheck;
 
 namespace Rubberduck.UI.Command
 {
@@ -10,11 +11,16 @@ namespace Rubberduck.UI.Command
     [ComVisible(false)]
     public class AboutCommand : CommandBase
     {
-        public AboutCommand() : base(LogManager.GetCurrentClassLogger()) { }
+        public AboutCommand(IVersionCheck versionService) : base(LogManager.GetCurrentClassLogger())
+        {
+            _versionService = versionService;
+        }
+
+        private readonly IVersionCheck _versionService;
 
         protected override void OnExecute(object parameter)
         {
-            using (var window = new AboutDialog())
+            using (var window = new AboutDialog(_versionService))
             {
                 window.ShowDialog();
             }
