+++ Rubberduck.Main/Rubberduck.Main.csproj
@@ -266,6 +266,7 @@
     <Compile Include="Root\RootComWrapperFactory.cs" />
     <Compile Include="Root\RubberduckIoCInstaller.cs" />
     <Compile Include="Root\TypeExtensions.cs" />
+    <Compile Include="VersionCheck.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="NLog.dll.nlog">
@@ -313,6 +314,9 @@
     <Reference Include="extensibility, Version=7.0.3300.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
       <EmbedInteropTypes>True</EmbedInteropTypes>
     </Reference>
+    <Reference Include="HtmlAgilityPack, Version=1.8.4.0, Culture=neutral, PublicKeyToken=bd319b19eaf3b43a, processorArchitecture=MSIL">
+      <HintPath>..\packages\HtmlAgilityPack.1.8.4\lib\Net45\HtmlAgilityPack.dll</HintPath>
+    </Reference>
     <Reference Include="Infralution.Localization.Wpf">
       <HintPath>..\libs\Infralution.Localization.Wpf.dll</HintPath>
     </Reference>
@@ -330,6 +334,7 @@
     <Reference Include="System.Data" />
     <Reference Include="System.Drawing" />
     <Reference Include="System.IO.Compression" />
+    <Reference Include="System.Net.Http" />
     <Reference Include="System.Runtime.Remoting" />
     <Reference Include="System.Runtime.Serialization" />
     <Reference Include="System.ServiceModel" />
