+++ Rubberduck.Main/packages.config
@@ -4,6 +4,7 @@
   <package id="Castle.Core" version="4.2.1" targetFramework="net46" />
   <package id="Castle.Windsor" version="4.1.0" targetFramework="net46" />
   <package id="EasyHook" version="2.7.6684" targetFramework="net46" />
+  <package id="HtmlAgilityPack" version="1.8.4" targetFramework="net46" />
   <package id="NLog" version="4.5.4" targetFramework="net46" />
   <package id="NLog.Schema" version="4.5.4" targetFramework="net46" />
 </packages>
